+++ settings.gradle.kts
@@ -67,7 +67,7 @@ include(
     ":subprojects:analysis-kotlin-descriptors:compiler",
     ":subprojects:analysis-kotlin-descriptors:ide",
     ":subprojects:analysis-kotlin-symbols",
-    ":subprojects:analysis-kotlin-symbols:compiler",
+    ":subprojects:analysis-kotlin-symbols:plugin",
     ":subprojects:analysis-markdown-jb",
 
     ":runners:gradle-plugin",
