--- subprojects/analysis-kotlin-symbols/compiler/src/main/resources/META-INF/services/org.jetbrains.dokka.plugability.DokkaPlugin
@@ -1,1 +1,0 @@
-org.jetbrains.dokka.analysis.kotlin.symbols.compiler.SymbolsAnalysisPlugin
