+++ subprojects/analysis-kotlin-symbols/plugin/api/compiler.api
