+++ subprojects/analysis-kotlin-symbols/plugin/build.gradle.kts
