+++ subprojects/analysis-kotlin-symbols/plugin/src/main/resources/META-INF/services/org.jetbrains.dokka.plugability.DokkaPlugin
@@ -1,0 +1,1 @@
+org.jetbrains.dokka.analysis.kotlin.symbols.plugin.SymbolsAnalysisPlugin
